simplified the DiffUNet architecture, and modified the diffDPIR and DPS demos to use the noise level sigma instead of the timestep.
